Fix quiz result page idempotency by moving QuizAttempt creation to final answer action

+++ app/controllers/quiz_controller.rb
@@ -10,6 +10,8 @@ class QuizController < ApplicationController
       end
     else
       session[:quiz] = { "module_id" => @module.id, "question_index" => 0, "score" => 0 }
+      session.delete(:last_attempt_id)
+      session.delete(:last_score)
     end
 
     @current_index = session[:quiz]["question_index"]
@@ -51,19 +53,41 @@ class QuizController < ApplicationController
     @next_index = @current_index + 1
     session[:quiz]["question_index"] = @next_index
     @next_question = questions[@next_index]
+
+    if @next_question.nil?
+      score = session[:quiz]["score"]
+      if current_user
+        attempt = QuizAttempt.create!(user: current_user, quiz_module: @module, score: score)
+        session[:last_attempt_id] = attempt.id
+      else
+        session[:last_score] = { "score" => score, "total" => @total_questions }
+      end
+      session.delete(:quiz)
+    end
+
     @options = @question.options.order(:id)
     render :answer
   end
 
   def result
-    score = session.dig(:quiz, "score") || 0
-    total = @module.questions.count
     if current_user
-      QuizAttempt.create(user: current_user, quiz_module: @module, score: score)
+      attempt = current_user.quiz_attempts.find_by(id: session[:last_attempt_id]) ||
+                current_user.quiz_attempts.where(quiz_module: @module).last
+      if attempt
+        @score = attempt.score
+        @total = @module.questions.count
+      else
+        redirect_to play_quiz_module_path(@module.slug, locale: I18n.locale) and return
+      end
+    else
+      last_score = session[:last_score]
+      if last_score
+        @score = last_score["score"]
+        @total = last_score["total"]
+      else
+        redirect_to play_quiz_module_path(@module.slug, locale: I18n.locale) and return
+      end
     end
-    @score = score
-    @total = total
-    session.delete(:quiz)
   end
 
   private
